Fix kubernetes dns
In kubernetes if we do not set eureka.instance.hostname with service name then eureka take host name as container name which has been created by kubernetes replication controller. Since eureka and fastpass service are deployed in different node so when fastpass consol try to call fasspass service it try to access with container name which is not reachable due to it deployed in different node. 

To overcome this issue eureka.instance.hostname name needs to set with kubernetes service name. Once it is done then kubernetes dns able to find the location of service and the service responsibility is to connect with appropriate pod.